fix merge issues

+++ lifelong_rl/lifelong_rl/trainers/q_learning/sac.py
@@ -114,11 +114,12 @@ class SACTrainer(TorchTrainer):
                                                                    return_log_prob=True)
             new_obs_actions = torch.zeros((len(new_obs_actions_num), self.env.action_space.n), dtype=torch.float32,
                                           device=new_obs_actions_num.device)
-            new_obs_actions.scatter_(1, new_obs_actions_num.unsqueeze(1), 1)
+            new_obs_actions.scatter_(1, new_obs_actions_num.unsqueeze(1).long(), 1)
+
 
             new_next_actions = torch.zeros((len(new_next_actions_num), self.env.action_space.n), dtype=torch.float32,
                                           device=new_next_actions_num.device)
-            new_next_actions.scatter_(1, new_next_actions_num.unsqueeze(1), 1)
+            new_next_actions.scatter_(1, new_next_actions_num.unsqueeze(1).long(), 1)
 
             # new_obs_actions = new_obs_actions.unsqueeze(1).float()
             # new_next_actions = new_next_actions.unsqueeze(1).float()
